@@ -2,13 +2,10 @@ name: child_builder
 description: A very simple builder that can be used by UI frameworks to customize / wrap internal widgets.
 homepage: https://github.com/peiffer-innovations/child_builder
 
-version: 1.0.0+2
+version: 1.1.0-nullsafety
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-
-analyzer:
-  exclude: [lib/generated/**]
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   flutter:
